+++ app/src/main/java/com/example/restcountriesapp/MainActivity.kt
@@ -8,12 +8,13 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
+import com.example.restcountriesapp.app.AppContainer
 import com.example.restcountriesapp.ui.theme.RestCountriesAppTheme
-import com.example.restcountriesapp.data.CountriesRepository
 import com.example.restcountriesapp.home.HomeRoute
 
 class MainActivity : ComponentActivity() {
-    private val countriesRepository = CountriesRepository()
+
+    private val appContainer = AppContainer()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -25,7 +26,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize()
                 ) { innerPadding ->
                     HomeRoute(
-                        countriesRepository = countriesRepository,
+                        countriesRepository = appContainer.countryRepository,
                         modifier = Modifier.padding(innerPadding)
                     )
                 }