+++ app/src/main/java/com/example/restcountriesapp/data/remote/RestCountriesApi.kt
@@ -1,0 +1,13 @@
+package com.example.restcountriesapp.data.remote
+
+import com.example.restcountriesapp.data.remote.dto.CountriesResponseDto
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface RestCountriesApi {
+
+    @GET("countries/v5")
+    suspend fun getCountries(
+        @Query("limit") limit: Int = 100
+    ): CountriesResponseDto
+}