+++ app/src/main/java/com/example/restcountriesapp/domain/model/Country.kt
@@ -1,0 +1,13 @@
+package com.example.restcountriesapp.domain.model
+
+data class Country(
+    val name: String,
+    val capital: String,
+    val region: String,
+    val population: Long,
+    val code: String,
+    val flagUrl: String?,
+    val latitude: Double?,
+    val longitude: Double?
+
+)